bal-women/requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
Pillow==9.2.0
psycopg2-binary==2.9.5 
gunicorn==20.1.0
pytz==2022.5
sqlparse==0.4.3
tzdata==2022.5

